+++ resources/views/app/activity-log/description-badge.blade.php
@@ -1,0 +1,1 @@
+<span class="badge badge-pill {{ $badgeClass }} px-3 py-2">{{ $label }}</span>